refactor(app): remove compound component
